# openapi.yaml
openapi: 3.0.3
info:
  title: CaseCellShop API
  description: |
    API backend da CaseCellShop - catálogo de produtos com cache,
    checkout assíncrono e consulta de status de pedidos.
  version: 1.0.0
  contact:
    name: CaseCellShop Dev Team

servers:
  - url: http://localhost:3000
    description: Servidor local de desenvolvimento

paths:
  /products:
    get:
      summary: Lista todos os produtos
      description: |
        Retorna catálogo de produtos com cache (TTL 30s, stale-while-revalidate).
        Cache hit retorna dados em <5ms. Cache miss consulta repositório (~100ms).
      operationId: listProducts
      tags: [Products]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Lista de produtos
          headers:
            X-Correlation-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer
                      correlationId:
                        type: string
        '500':
          $ref: '#/components/responses/InternalError'

  /products/{productId}:
    get:
      summary: Busca produto por ID
      operationId: getProduct
      tags: [Products]
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Produto encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /checkout:
    post:
      summary: Inicia checkout assíncrono
      description: |
        Cria um pedido, reserva estoque atomicamente e enfileira para processamento
        pelo ERP. Retorna 202 Accepted com orderId para acompanhamento.
        
        Suporta idempotência via idempotencyKey - requisições duplicadas retornam
        o mesmo pedido sem efeitos colaterais.
      operationId: createCheckout
      tags: [Checkout]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
            examples:
              singleItem:
                summary: Checkout com um item
                value:
                  items:
                    - productId: "prod-001"
                      quantity: 2
                  idempotencyKey: "client-uuid-123"
              multipleItems:
                summary: Checkout com múltiplos itens
                value:
                  items:
                    - productId: "prod-001"
                      quantity: 1
                    - productId: "prod-002"
                      quantity: 3
                  idempotencyKey: "client-uuid-456"
      responses:
        '202':
          description: Pedido aceito para processamento
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/CheckoutResponse'
                  meta:
                    type: object
                    properties:
                      correlationId:
                        type: string
        '400':
          description: Request inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missingItems:
                  value:
                    error:
                      code: INVALID_REQUEST
                      message: "items é obrigatório e deve ser um array não vazio"
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Estoque insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                outOfStock:
                  value:
                    error:
                      code: INSUFFICIENT_STOCK
                      message: "Estoque insuficiente para produto: prod-003"
        '500':
          $ref: '#/components/responses/InternalError'

  /orders/{orderId}/status:
    get:
      summary: Consulta status do pedido
      description: |
        Retorna o status atual do pedido. Status possíveis:
        - pending: pedido criado, aguardando processamento
        - processing: worker está processando com o ERP
        - confirmed: ERP confirmou o faturamento
        - failed: falha após máximo de retries
      operationId: getOrderStatus
      tags: [Orders]
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Status do pedido
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/OrderStatus'
                  meta:
                    type: object
                    properties:
                      correlationId:
                        type: string
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /health:
    get:
      summary: Health check
      operationId: healthCheck
      tags: [System]
      responses:
        '200':
          description: Serviço saudável
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [healthy]
                  timestamp:
                    type: string
                    format: date-time
                  queueSize:
                    type: integer

  /admin/reconcile:
    post:
      summary: Trigger manual de reconciliação
      description: |
        Executa reconciliação manualmente: republica pedidos "pending" antigos na fila
        e marca pedidos "processing" stuck como failed (com rollback de estoque).
      operationId: triggerReconciliation
      tags: [Admin]
      responses:
        '200':
          description: Resultado da reconciliação
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      republished:
                        type: integer
                        description: Pedidos pending republicados na fila
                      failed:
                        type: integer
                        description: Pedidos stuck marcados como failed
                      errors:
                        type: array
                        items:
                          type: string

  /metrics:
    get:
      summary: Métricas Prometheus
      description: Expõe métricas no formato Prometheus para scraping
      operationId: getMetrics
      tags: [System]
      responses:
        '200':
          description: Métricas em formato text/plain
          content:
            text/plain:
              schema:
                type: string

components:
  parameters:
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: ID de correlação para rastreabilidade. Gerado automaticamente se não fornecido.
      schema:
        type: string
        format: uuid

  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
          example: "prod-001"
        name:
          type: string
          example: "Capinha iPhone 15 Pro - Transparente"
        price:
          type: number
          format: float
          example: 49.90
        stock:
          type: integer
          example: 100
        category:
          type: string
          example: "iphone"
        imageUrl:
          type: string
          example: "/images/iphone15-clear.jpg"
      required: [id, name, price, stock, category]

    CheckoutRequest:
      type: object
      properties:
        items:
          type: array
          minItems: 1
          items:
            type: object
            properties:
              productId:
                type: string
              quantity:
                type: integer
                minimum: 1
            required: [productId, quantity]
        idempotencyKey:
          type: string
          description: Chave única para garantir idempotência (UUID recomendado)
      required: [items, idempotencyKey]

    CheckoutResponse:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, processing, confirmed, failed]
        message:
          type: string

    OrderStatus:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, processing, confirmed, failed, cancelled]
        totalAmount:
          type: number
          format: float
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
              quantity:
                type: integer
              unitPrice:
                type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        retryCount:
          type: integer
        errorMessage:
          type: string
          nullable: true

    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
        meta:
          type: object
          properties:
            correlationId:
              type: string

  responses:
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Erro interno do servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'